+++ app/stores/SidebarStore.js
@@ -12,11 +12,10 @@ export default class SidebarStore extends Store {
     SETTINGS: ROUTES.SETTINGS.ROOT,
   };
 
-  INITIAL_HIDE_SUB_MENU_DELAY = 4000;
-  ACTION_HIDE_SUB_MENU_DELAY = 1000;
+  DEFAULT_WINDOW_WIDTH = 1150;
 
   @observable activeSidebarCategory: string = this.CATEGORIES.WALLETS;
-  @observable isShowingSubMenus: bool = true;
+  @observable isShowingSubMenus: bool = false;
 
   _hideSubMenuTimeout = null;
 
@@ -25,11 +24,9 @@ export default class SidebarStore extends Store {
     actions.toggleSubMenus.listen(this._toggleSubMenus);
     actions.activateSidebarCategory.listen(this._onActivateSidebarCategory);
     actions.walletSelected.listen(this._onWalletSelected);
-    this.actions.networkStatus.isSyncedAndReady.listen(() => {
-      this._hideSubMenusAfterDelay(this.INITIAL_HIDE_SUB_MENU_DELAY);
-    });
     this.registerReactions([
       this._syncSidebarRouteWithRouter,
+      this._setInitialSubMenusState,
     ]);
   }
 
@@ -45,18 +42,12 @@ export default class SidebarStore extends Store {
 
   @action _toggleSubMenus = () => {
     this.isShowingSubMenus = !this.isShowingSubMenus;
-    this._clearExistingHideSubMenuTimeout();
-  };
-
-  @action _hideSubMenus = () => {
-    this.isShowingSubMenus = false;
-    this._clearExistingHideSubMenuTimeout();
+    window.removeEventListener('resize', this._handleWindowResize);
   };
 
   @action _onActivateSidebarCategory = (
       { category, showSubMenus }: { category: string, showSubMenus: boolean }
     ) => {
-    this._clearExistingHideSubMenuTimeout();
     if (category !== this.activeSidebarCategory) {
       this.activeSidebarCategory = category;
       if (showSubMenus != null) this.isShowingSubMenus = showSubMenus;
@@ -69,18 +60,27 @@ export default class SidebarStore extends Store {
     }
   };
 
-  @action _hideSubMenusAfterDelay = (delay: number) => {
-    this._clearExistingHideSubMenuTimeout();
-    this._hideSubMenuTimeout = setTimeout(this._hideSubMenus, delay);
+  @action _handleWindowResize = () => {
+    if (document.documentElement.clientWidth < this.DEFAULT_WINDOW_WIDTH) {
+      this.isShowingSubMenus = false;
+    } else {
+      this.isShowingSubMenus = true;
+    }
   };
 
   @action _onWalletSelected = ({ walletId }: { walletId: string }) => {
     this.stores.wallets.goToWalletRoute(walletId);
-    this._hideSubMenusAfterDelay(this.ACTION_HIDE_SUB_MENU_DELAY);
   };
 
   @action _setActivateSidebarCategory = (category: string) => {
     this.activeSidebarCategory = category;
+  };
+
+  _setInitialSubMenusState = () => {
+    if (document.documentElement.clientWidth >= this.DEFAULT_WINDOW_WIDTH) {
+      this.isShowingSubMenus = true;
+    }
+    window.addEventListener('resize', this._handleWindowResize);
   };
 
   _syncSidebarRouteWithRouter = () => {
@@ -91,10 +91,6 @@ export default class SidebarStore extends Store {
       if (route.indexOf(category) === 0) this._setActivateSidebarCategory(category);
     });
   };
-
-  _clearExistingHideSubMenuTimeout() {
-    if (this._hideSubMenuTimeout) clearTimeout(this._hideSubMenuTimeout);
-  }
 
 }